Don't tile tilix --quake window
Some people (like me) use the drop down tilix --quake window for a quick
terminal session. Don't try to tile this window. Seems like the only way
to tell the difference between a normal tilix window and the quake
window is the window role[1]. Check a window's role and compare it to
the value 'quake'.

[1]: https://github.com/gnunn1/tilix/issues/619