Issue #3: Added window::is_ready function
This function will be used to check if a window is successfully created
and ready. It checks if the window size is zero. A zero sized window
will not be used in a layout. This leeds to a bug that a newly created
window on wayland is not moved immediately into its layout. Some sort
of timer is required.